Project/requirements.txt
# Core AI/ML Framework
agno>=0.1.0
google-generativeai>=0.8.0

# Web & API Frameworks
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Data & Visualization
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.17.0

# Database & ORM
sqlalchemy>=2.0.0

# Environment & Configuration
python-dotenv>=1.0.0
pyyaml>=6.0

# Logging
python-json-logger>=2.0.7

# Date/Time Handling
python-dateutil>=2.8.2

# HTTP Clients
requests>=2.31.0
httpx>=0.25.0

# CLI Interface
click>=8.1.0
tabulate>=0.9.0
colorama>=0.4.6

# System Monitoring
psutil>=5.9.0

# Code Quality & Testing
pytest>=7.4.0
pytest-html>=4.0.0
pytest-cov>=4.1.0
black>=23.10.0
flake8>=6.1.0
